+++ meridian/analysis/visualizer_test.py
@@ -1514,6 +1514,41 @@ class MediaSummaryTest(parameterized.TestCase):
     self.media_summary_kpi = visualizer.MediaSummary(meridian_kpi)
     self.media_summary_revenue_2 = visualizer.MediaSummary(meridian_revenue_2)
 
+  def test_media_summary_init_non_media_baseline_values(self):
+    non_media_baseline_values = ["min", "max"]
+    media_summary = visualizer.MediaSummary(
+        self.meridian_revenue,
+        non_media_baseline_values=non_media_baseline_values,
+    )
+    self.assertEqual(
+        media_summary._non_media_baseline_values, non_media_baseline_values
+    )
+
+  def test_media_summary_all_summary_metrics_calls_analyzer_correctly(self):
+    non_media_baseline_values = [1.0, "max"]
+    media_summary = visualizer.MediaSummary(
+        self.meridian_revenue,
+        non_media_baseline_values=non_media_baseline_values,
+    )
+    _ = media_summary.all_summary_metrics
+    self.mock_analyzer_summary_metrics.assert_called_with(
+        selected_times=None,
+        use_kpi=False,
+        confidence_level=0.9,
+        include_non_paid_channels=True,
+        non_media_baseline_values=non_media_baseline_values,
+    )
+
+  def test_media_summary_update_non_media_baseline_values(self):
+    media_summary = visualizer.MediaSummary(self.meridian_revenue)
+    non_media_baseline_values = [1.0, "max"]
+    media_summary.update_summary_metrics(
+        non_media_baseline_values=non_media_baseline_values
+    )
+    self.assertEqual(
+        media_summary._non_media_baseline_values, non_media_baseline_values
+    )
+
   def test_media_summary_plot_contribution_waterfall_different_scenarios(self):
     # Verifies each of the three kpi_type and revenue_per_kpi scenarios have
     # the correct labels associated with them.